requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
dj-database-url==0.5.0
Django==2.1.5
django-heroku==0.3.1
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.8
psycopg2==2.8.5
pytz==2019.3
requests==2.22.0
urllib3==1.25.8
whitenoise==5.0.1
